Patched kibana configs to support ssl
Previously kibana configs forced http:// for
the API calls it makes, which cause lots of angry browser
errors when using SSL. This patch allows it to use the
protocol that was used when requesting the page, for greater
fixiness.

This is a blatant rip off from James Hunt (@filefrog)'s commit
on Long Nguyen's logstash docker image repo.